Remove Medium link from footer
We only make a couple Medium posts a year, and the platform has fallen off significantly as a site people actively check for new posts from organizations they follow. We should definitely highlight our Medium posts on the News page when we publish them, but a handful of new posts a year doesn't justify taking up space this valuable.